refactor(python): share Gemini usage normalization
